# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org.quicko.qpon",
    "description": "Qpon is a coupon framework built by Quicko. Built in a scalable, multi-tenant thought process, the Qpon framework enables businesses to do campaigns and run offers using the Qpon API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{quicko_host}}"
    }
  ],
...
  "paths": {
...
    "/coupons/{coupon_id}/campaigns": {
      "post": {
        "tags": [
          "Campaign"
        ],
        "summary": "Create campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "campaign"
                  },
                  "name": {
                    "type": "string"
                  },
                  "budget": {
                    "type": "number"
                  },
                  "external_id": {
                    "type": "string"
                  },
                  "campaign-status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "exhausted"
                    ]
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Fetch campaigns",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupon sets by specified status",
            "example": "active"
          },
          {
            "name": "budgeted",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Optional filter to get coupon sets with budget",
            "example": "true"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "4a4e1a61-0c9c-4941-9b83-534d1fe78226"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{coupon_id}/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Fetch campaign",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Campaign"
        ],
        "summary": "Update campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "@entity": "org.quicko.qpon.campaign",
                  "id": "{{campaign_id}}",
                  "name": "AY 2024-2025 ITR Filing"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaign"
        ],
        "summary": "Delete campaign",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{coupon_id}/campaigns/{campaign_id}/deactivate": {
      "post": {
        "tags": [
          "Campaign"
        ],
        "summary": "Deactivate campaign",
        "requestBody": {
          "content": {}
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{coupon_id}/campaigns/{campaign_id}/reactivate": {
      "post": {
        "tags": [
          "Campaign"
        ],
        "summary": "Reactivate campaign",
        "requestBody": {
          "content": {}
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/summary": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Fetch campaign summary",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "campaign_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "{{campaign_id}}"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}